requirements.txt
numpy==1.24.2
psycopg2-binary==2.9.5
PyYAML==6.0
pyodbc==4.0.35
mysql-connector-python==8.0.32
PyMySQL==1.0.3